requirements.txt
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.1.3
tabulate>=0.9.0

# Optional but recommended for more file format support
PyPDF2>=3.0.0          # For PDF files
python-docx>=1.0.0     # For DOCX files
striprtf>=0.0.22       # For RTF files

# Optional: uncomment and install for more accurate NLP-based detection
# spacy>=3.7.2
